full code committed

+++ mycontacts-backend/controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all Contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   // res.status(200).json({ message: "Get all Contacts" });
   res.status(200).json(contacts);
 });
 
 //@desc Create new Contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is : ", req.body);
   const { name, email, phone } = req.body;
@@ -25,6 +25,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   // res.status(201).json({ message: "Create Contact" });
   res.status(201).json(contact);
@@ -32,7 +33,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get all Contacts
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -46,26 +47,43 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update Contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
-  const contact = await Contact.findById(req.params.id);
-  if (!contact) {
-    res.status(404);
-    throw new Error("Contact not found");
-  }
+  const contactId = req.params.id;
 
-  const updatedContact = await Contact.findByIdAndUpdate(
-    req.params.id,
-    req.body,
-    { new: true }
-  );
-  // res.status(200).json({ message: `Update Contact for ${req.params.id}` });
-  res.status(200).json(updatedContact);
+  try {
+    const contact = await Contact.findById(contactId);
+
+    if (!contact) {
+      res.status(404);
+      throw new Error("Contact not found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+      res.status(403);
+      throw new Error(
+        "User doesn't have permission to update other user's contacts"
+      );
+    }
+
+    const updatedContact = await Contact.findByIdAndUpdate(
+      contactId,
+      req.body,
+      { new: true }
+    );
+
+    res.status(200).json({ message: `Contact with ID ${contactId} has been updated`, updatedContact });
+
+  } catch (error) {
+    console.error("Error updating contact:", error);
+    res.status(500).json({ message: "An error occurred while updating the contact" });
+  }
 });
+
 
 //@desc Delete Contacts
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   // const contact = await Contact.findById(req.params.id);
   const contactId = req.params.id;
@@ -75,7 +93,14 @@ const deleteContact = asyncHandler(async (req, res) => {
       res.status(404);
       throw new Error("Contact not found");
     }
-    console.log("Contact found:", contact);
+
+    if (contact.user_id.toString() !== req.user.id) {
+      res.status(403);
+      throw new Error(
+        "User doesn't have permission to delete other user's contacts"
+      );
+    }
+
 
     await Contact.findByIdAndDelete(contactId);
     console.log("Contact successfully deleted");
